Add .env file to .gitignore and create bootstraping/azure.bash script (#6)
This commit adds the .env file to the .gitignore to prevent it from being tracked in the repository. It also introduces the bootstraping/azure.bash script, which is used for setting up an Azure environment. The script includes functionality for loading the GitHub token from the .env file, parsing command line arguments, checking login status for GitHub and Azure, creating resource groups and service principals, and saving values in the terraform.tfvars file. Additionally, it updates the repo actions secrets and variables with the necessary values.